Refactor TLS credential and session handling
TLS Credentials and Sessions now have a thin base class which
Connections know about, details are left to child classes.  The user
must construct a shared_ptr<TLSCreds> to give to Client/Server on
construction.  The user MAY keep their shared_ptr if they want to use
the same credentials object for multiple Clients/Servers.  When a
Client/Server creates a Connection, the Connection gets a copy of the
shared_ptr<TLSCreds> and uses it to make a unique_ptr<TLSSession>.

Currently only GNUTLS is implemented, but "no crypto" should be fairly
trivial to implement.  The TLSCreds and TLSSession interface is subject
to change of course if other TLS methods/libs need something the current
interface cannot provide.

Unit tests updated and passing.