gnutls from-pubkey key verify and ALPN check
GNUTLSCreds now supports a key verify function in raw pubkey mode

clients send the client ALPN during negotiation,
relays send the relay ALPN on outbound, and
relays set both for inbound.
This is used to determine whether we are being connected to by
a relay or a client, and check keys / setup lokinet accordingly.